refactor(cli): A.3.5 业务 inline eval gold 3 sub + 删 src/cli/commands directory
eval-gold.ts 3 个 sub-sub handler（init / validate / compare）业务整段 inline
到 src/cli/oclif/commands/eval/gold/{init,validate,compare}.ts:run() body。
删 src/cli/commands/eval-gold.ts。

src/cli/commands/ directory 整个删除（8 个业务文件全部 inline 完成，_shared.ts
之前已上移到 src/cli/）。所有 src/cli/commands/ import 0 处，100% oclif-native
的「CLI entry 全走 oclif，业务不再住独立 commands directory」目标达成。

yarn lint && build && test 1494 全绿。
